Add job queue management
