+++ root/cron.php
@@ -14,6 +14,7 @@
 
 require_once __DIR__ . '/config.php';
 require_once __DIR__ . '/autoload.php';
+require_once __DIR__ . '/jobs.php';
 
 use App\Core\ErrorMiddleware;
 use App\Controllers\StatusController;
@@ -21,6 +22,7 @@ use App\Models\Account;
 use App\Models\User;
 use App\Models\Feed;
 use App\Models\Database;
+use App\Models\JobQueue;
 use App\Core\Utility;
 
 // Apply configured runtime limits after loading settings
@@ -322,65 +324,6 @@ function clearList(): bool
     return true;
 }
 
-/**
- * Populate the next 24-hour job schedule into status_jobs.
- */
-function fillQueryJobs(): bool
-{
-    global $debugMode;
-    $now = new DateTime();
-    $end = (clone $now)->modify('+24 hours');
-    logDebug('Generating job queue up to ' . $end->format('Y-m-d H:i:s'));
-
-    $accounts = Account::getAllAccounts();
-    if (empty($accounts)) {
-        logDebug('No accounts found to queue.');
-        return true;
-    }
-
-    foreach ($accounts as $account) {
-        $hours = array_filter(array_map('trim', explode(',', $account->cron)), 'strlen');
-        if (empty($hours)) {
-            continue;
-        }
-
-        $days = array_map('strtolower', array_map('trim', explode(',', $account->days)));
-        foreach ($hours as $hour) {
-            if (!is_numeric($hour)) {
-                continue;
-            }
-            $runTime = new DateTime();
-            $runTime->setTime((int) $hour, 0);
-            if ($runTime < $now) {
-                $runTime->modify('+1 day');
-            }
-            if ($runTime > $end) {
-                continue;
-            }
-            $dayName = strtolower($runTime->format('l'));
-            if (!in_array('everyday', $days) && !in_array($dayName, $days)) {
-                continue;
-            }
-
-            $runAt = $runTime->format('Y-m-d H:i:s');
-            $db = new Database();
-            $db->query('SELECT id FROM status_jobs WHERE username = :u AND account = :a AND run_at = :r LIMIT 1');
-            $db->bind(':u', $account->username);
-            $db->bind(':a', $account->account);
-            $db->bind(':r', $runAt);
-            if (!$db->single()) {
-                $db->query("INSERT INTO status_jobs (username, account, run_at, status) VALUES (:u, :a, :r, 'pending')");
-                $db->bind(':u', $account->username);
-                $db->bind(':a', $account->account);
-                $db->bind(':r', $runAt);
-                $db->execute();
-                logDebug('Queued job for ' . $account->account . ' at ' . $runAt);
-            }
-        }
-    }
-
-    return true;
-}
 
 /**
  * Run queued jobs and generate statuses.
@@ -389,10 +332,8 @@ function runQueuedJobs(): bool
 {
     global $debugMode;
     $limit = defined('CRON_QUEUE_LIMIT') ? (int) CRON_QUEUE_LIMIT : 10;
+    $jobs = JobQueue::getPending($limit);
     $db = new Database();
-    $db->query("SELECT * FROM status_jobs WHERE status = 'pending' AND run_at <= NOW() ORDER BY run_at ASC LIMIT :l");
-    $db->bind(':l', $limit, PDO::PARAM_INT);
-    $jobs = $db->resultSet();
 
     if (empty($jobs)) {
         logDebug('No queued jobs to run.');
